Getting latest changes from factory fix onto my laptop

+++ db/migrate/20140724024644_create_locations.rb
@@ -1,8 +1,8 @@
 class CreateLocations < ActiveRecord::Migration
   def change
     create_table :locations do |t|
-      t.integer :latitude
-      t.integer :longitude
+      t.decimal :latitude
+      t.decimal :longitude
       t.string :name
       t.string :venue_type
       t.string :address
